messenger/requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
boto3==1.10.29
botocore==1.13.29
coverage==5.0.3
decorator==4.4.1
Django==2.2.5
django-cors-headers==3.2.0
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.0.1
freeze==1.0.10
gunicorn==19.9.0
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
jmespath==0.9.4
more-itertools==7.2.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.0
prompt-toolkit==2.0.10
psycopg2-binary==2.8.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.2
pytest==5.2.1
python-dateutil==2.8.0
python-magic==0.4.15
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
